refactor: 지원하는 OAuth 제공자 관리 관련 코드 분리 #89
